split staff.js commands
