+++ Panel/bot/discord/events/message.js
@@ -69,7 +69,7 @@ module.exports = (client, message) => {
 
         if (blacklisted.includes(message.channel.id) && (message.member.roles.cache.find(x => x.id === '748117822370086932') == null) &&
             !(message.channel.id === '738548111323955270' && command === 'info')) return;
-        if (command === "server" || command == "user" || command == "staff") {
+        if (command === "server" || command === "user" || command === "staff") {
             //Cooldown setting
             let cooldown = {};
